Refactor autoProcessMusic -> music
