+++ core/auto_process/music.py
