+++ nzbToMedia.py
@@ -632,7 +632,7 @@ from core import logger, nzbToMediaDB
 from core.auto_process.comics import Comic
 from core.auto_process.autoProcessGames import Game
 from core.auto_process.autoProcessMovie import Movie
-from core.auto_process.autoProcessMusic import Music
+from core.auto_process.music import Music
 from core.auto_process.tv import TV
 from core.nzbToMediaUserScript import external_script
 from core.nzbToMediaUtil import char_replace, clean_dir, convert_to_ascii, extract_files, get_dirs, get_download_info, get_nzoid, plex_update, update_download_info_status
